Multi-select support for opening PuTTY sessions to multiple devices.  Also added feature support for changing the window title of the PuTTY session so that windows can more easily be identified when switch between windows with Alt-Tab.
The PuTTY module now requires that host information for the device that you
want to connect to is passed using PSCustomObject with two properties:
Hostname, and IP. This is to provide a standard interface for passing data
to the module, and puts the responsibility on the caller to understand and
properly format the information.

To allow for changing the title of the PuTTY window, sessions are created
from the default.reg file with updated HostName and WinTitle settings. The
session configuration is then merged into the registry under CurrentUser,
where the PuTTY application stores it's different configuration settings
for session profiles.  The session profiles are named after the device
hostname.  The only allowed special character is '-'.  All other non-word
characters are converted into a '-' in order to conform the hostname with
windows registry rules.