[SortedTreeView] Debug - GroupStaticPanel add grouping level button
New-AddGroupButton : The term 'New-AddGroupButton' is not recognized as the name of a cmdlet, function, script file, or operable program. Check the spelling of the name, or if a path was included, verify that the path is correct and try again.
At C:\Users\Danny\Documents\WindowsPowerShell\Programs\NOSC\modules\SortedTreeView\SortedTreeView.psm1:940 char:15
+     $Button = New-AddGroupButton @ButtonParams
+               ~~~~~~~~~~~~~~~~~~
    + CategoryInfo          : ObjectNotFound: (New-AddGroupButton:String) [], CommandNotFoundException
    + FullyQualifiedErrorId : CommandNotFoundException

The static button for adding new setting panels had been refactored into a
common function, New-SettingStaticButton, but the call was never updated in
the New-GroupStaticPanel function.