[Device/SingleView] Debug - Object initialization errors
The property 'DataView' cannot be found on this object. Verify that the property exists and can be set.
At C:\Users\Danny\Documents\WindowsPowerShell\Programs\NOSC\modules\Device\SingleView.psm1:359 char:9
+         $DataNodeDefinition.Custom.DataView = $DataView
+         ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
    + CategoryInfo          : InvalidOperation: (:) [], RuntimeException
    + FullyQualifiedErrorId : PropertyNotFound

Updating the DataNodeDefinition should be completed before calling the
intitialization function for the SortedTreeView component.  Additionally
the ViewControl function was missing a call to Controls.Add() for the
TreeView as SortedTreeView no longer adds itself directly to a control, but
expects the caller to perform that task.