[SortedTreeView] Refactor - SettingsManager
SettingsManager is a new object that combines the scattered references for
the settings collections and other important control references and methods.
This consolidation allows for the SettingsManager to be the one stop shop for
outside controls/callers to be able to check the status of the SortedTreeView
settings and apply them as necessary.

Also, this design eliminates the need to "Float" up the Apply button from the
static panel for the data node.  Overall, this should make it much easier to
see where this object comes from, and how the settings are manipulated from
outside of the module.