[SortedTreeView] Debug - Setting panel not properly initializing the field selector
Initialization of the combo box items isn't completed until after the object
has been added to the parent control and displayed.  As such, setting the
SelectedItem property is only effective after after the SelectedValueChanged
event has fired for the first time.

To ensure the proper initialization of the combo box a flag has been added
to keep track of the intiialization state.  After all of the dropdown items
for the control have been created and the SelectedValueChanged event is fired
for the first time, the flag is checked, and the SelectedItem updated to the
value set in the registration object during the creation of the setting panel.