[SortedTreeView] Debug - Leaf DataNode label text field not being properly registered
The registration of the field name to be used as the display text for DataNodes
was not being handled correctly after the refactoring.  Updated the logic
to make use of the SettingsCollection that has been attached to the static
DataNode panel, and some logic errors in the order of execution.