Debug Repository.Status().
Fixed the ConvertFrom-PSObject function in Common/Objects.psm1.

Made the Compare-Entry function a method of the Index object.

Fixed cTime and mTime property setting during entry creation within the
repository.

Added a reference for the content addressable file system manager object to
the index so that it can retrieve required objects and use the shared hash
routines.