Updated Index.psm1.
Fixed bugs where the CacheTree object type wasn't being set correctly.

Made the Index.TREE object a CacheTree so that it will work properly with
the Repository.Commit() and Index.Add() operations.

Updated the Index.Remove(Entry) operation.  Removed the parameter sets as
script methods don't support them for overloading.

Made sure that the root CacheTree was being properly invalidated when an
entry is added to the index.

Signed-off-by: alogical <alogical.2@gmail.com>