Updated Repository.Unstage(); Objects.psm1, Index.psm1, VersionControl.psm1
Fixed SHA1 hash string return from the ShaProvider.HashString(). The method
was not properly filtering extraneous output from the base object method
ComputeHash().

Added summarization support to the index CacheTree object.

Completed the Index.RevalidateTree(), Index.SummarizeTrees() methods.  These
methods allow the index to quickly identify if the index has been returned
to the original state of the HEAD commit in a deterministic manner.

Integrated the cache tree revalidation into the Repository.Unstage() method
to revert the index flags to show that the index is consistent with the
previously committed version of the working directory.