Bug fixes VersionControl.psm1, Index.psm1
Fixed bug where the idx.Entries array list was being overwritten with a
fixed size array during the Index.Checkout() routine.

Fixed bug in VersionControl.psm1 with Repository.Commit() where nested
directory structures paths weren't being properly reconstituted into the
proper relative path from the working directory by Build-Tree.

Refactored the Index.Checkout() operation to update the idx.HEAD and
idx.Commit values.  Removed the updates to these values from the
Repository.Commit() operation.

Updated the Index.Add() operation to allow for skipping the invalidation of
the TreeCache and writing the changes to disk.  This allows the Unstage()
operation to perform the rollback to the previously committed version, and
revalidate the CacheTree without duplicating work.

Updated the Get-Summary function and Index.SummarizeTrees() operations to
use a concatenated string of the file name and object ID in order to catch
name changes for index file entries.

Updated the name of the Index.RefreshCache() operation to RefreshPathCache
to better represent what is actually being refreshed.