Updated Commit() support.
Changed files:
  VersionControl.psm1
  Index.psm1

Fixed bug in Index where sub-trees weren't being properly built when a tree
had multiple sub-trees.

Updated the Commit() operation to save the root tree of the commit as a
separate object so that it can be used across multiple commits if needed.

Updated all of the operations that rely on the root tree of a commit to
load root tree object that is referenced by the commit object as it is no
longer directly within the commit.
  Index.Checkout()
  Repository.Unstage()