Updated VersionControl.psm1; finished Checkout() support.
Updated the loading of the index to properly show if the index has been
modified previously for the current repository.

Refactored GetHeadOid() into GetHeadName(), and GetHeadPath() for use within
other methods that need this information.

Added Repository.Modified script property which checks if the working
directory has been modified and returns true on the first modification
found, or false if there are no changes.

Added Repository.StageAll() to add all changes to the working directory at
once.

Updated Repository.LogBranch() to calculate the UtcTime and UtcOffset itself
instead of taking this data as parameters.

Finished the Repository.Checkout() support; .LogCheckout() and .Checkout().