Implementing OnClose logic for graphical interface and code cleanup.
Added OnClose handler to dispose of the SwisConnectionObject used by the
SolarWinds.psm1 module.  This is to ensure the TCP socket is closed in a
timely manner.

Fixed bug in SingleView with orion auth.  Settings .accessor was mistyped.

Refactored initialize-component logic, putting the responsibility of adding
the component controls to their parent control back onto the caller.